requirements.txt
Flask==2.3.3
requests>=2.31.0
azure-data-tables==12.4.4
pytz==2023.3
gunicorn==21.2.0
azure-functions==1.18.0